backend/requirements.txt
flask
python-dotenv
flask-cors
pytest-cov